+++ apps/search_engine/__init__.py
@@ -1,0 +1,12 @@
+# -*- encoding: utf-8 -*-
+"""
+Copyright (c) 2019 - present AppSeed.us
+"""
+
+from flask import Blueprint
+
+blueprint = Blueprint(
+    'searchengine_blueprint',
+    __name__,
+    url_prefix=''
+)
